Extraxted RatingView from Appetizer Page
