+++ Sources/HipMobileUI/Views/RatingView.xaml.cs
@@ -1,0 +1,14 @@
+﻿using PaderbornUniversity.SILab.Hip.Mobile.UI.Navigation;
+using PaderbornUniversity.SILab.Hip.Mobile.UI.ViewModels.Views;
+
+
+namespace PaderbornUniversity.SILab.Hip.Mobile.UI.Views
+{
+    public partial class RatingView : IViewFor<RatingViewModel>
+    {
+        public RatingView()
+        {
+            InitializeComponent();
+        }
+    }
+}
